Send converted document by email using Google Drive and Gmail.
The extension now sends the result from k2pdfopt to an email address
which is specified in a setting. Dontprint first uploads the file to the
user's Google Drive account using the Oauth2 mechanism to authorize.
Once the file is uploaded, the extension runs a small web app written
in Google Script that sends the uploaded file to a given email address.
The email address of the recipient is set by the extension during the upload
process by specifying some meta data on the file. This secures this
feture from being hijacked as a mass email sender since only people
who are authorized to set meta data on Google Drive files are able
to send them.

The following preference strings need to be set in firefox for this
extension to work (open "about:config" in firefox to set them):

  * extensions.zotero.dontprint.recipientEmail:
    The email address where the converted files are sent to.

  * extensions.zotero.dontprint.copyToMe:
    Boolean value; if set to true, the email goes Cc to the Gmail inbox.

  * extensions.zotero.dontprint.copyInGoogleDrive:
    Boolean value; if set to true, the uploaded file remains in
    Google Drive even after the Email is sent. If set to false,
    the file is marked as trashed in Google drive once the email
    is sent. (Deleting files permanently doesn't seem to be
    possible thru the API.)