Give feedback when user clicks "dontprint this page"
The "dontprint this page" icon changes into a progress icon when
clicked. When the user goes to a different page or tab, the progress
icon in the urlbar disappears (or is replaced by a "dontprint this
page" icon, if applicable for the new page/tab). When the user
comes back to the page where he clicked "dontprint this page" and
the page is still being processed then the icon changes back to
a progress icon. When the progress is done, the icon changes back
to a "dontprint this page" icon.

TODO: The tooltip of the progress icon states "click icon for
detailed progress information". This is not yet implemented.
Clicking on the progress icon should open a new tab with
a list of all currently running jobs and their status.