No longer use Google Drive.
Instead of using the detour to upload files first to Google Drive
and then from there sending them by email, Dontprint now just uploads
the file directly to the sendmail script. It uses a POST request
to send the data. Before sending the POST request, the Dontprint client
makes a dummy GET call to the same Google Apps Script application to
make sure that the subsequent POST request isn't interrupted by an
authorization page. The POST request is done with an XHR. An
alternative would have been to do send the POST request from the
already opened tab but that would freeze the tab (thus not allowing us
to show a progress bar) and change the tab title to "Connecting", which
is wrong.