Simplify control flow for Dontprint jobs
Previous to this commit, Dontprint jobs consisted of a series of
asynchroneous function calls, where each function initiated some
asynchroneous operation whose callback would be set to a function
that would initiate the next asynchroneous operation. This approach
created a very complex control flow and it was not obvious at first
sight, in which order operations were executed. This made error
handling difficult. Also tracking progress of a currently running job
would have been challenging within this framework because jobs didn't
have a direct representation in the data model, and it would have
been difficult to abort a running job.

This commit greately simplifies the complete control flow of a
Dontprint job:

 * Each job is now directly represented by an object, usually
   referenced to by a variable called "job", which holds all data
   associated with that job.

 * A single small function "runJob()" supervises the complete life
   cycle of a job and catches any errors.

 * The function runJob() spawns a Task.jsm task in which it iterates
   thru a number of asynchroneous operations. Each operation is
   implemented in a generator function so that runJob() can initiate
   the operation and wait for its completion with a single yield
   statement. If any of the sub-operations encounters an error,
   it can just throw an exception and it will be handled by runJob().