Remove all dependencies to Zotero.
Dontprint no longer requires Zotero to be installed. If Zotero
happens to be installed, Dontprint detects this at startup and
hooks into Zotero's web-site detection and translation code. If
Zotero isn't installed, then Dontprint programmatically loads
a (unfortunately, not quite so) minimal on-board version of
Zotero. The on-board version of Zotero is included in a git
submodule "zotero". This submodule points to a fork of the
official zotero code where some hard-coded references had to
be adjusted to work with Dontprint. The submodule should make
it easier to merge in future changes to Zotero.

Use the new bash script "install" to fetch all required code
from submodules.

Dontprint doesn't include any of the code in the "zotero"
directory directly. Only parts of the code are included via
some symlinks. When it comes to packaging Dontprint, I should
write a packaging script that automatically resolves these
symlinks and packages only the parts of Zotero that are
really used by Dontprint.