Add first post translator (arXiv)
Zotero's translators are a bit conservative for assumptions on articles
that come from preprint servers. This is arguably good for a literature
database such as Zotero, but it means that one can make no assumptions
on margins for articles from, e.g., arXiv. This commit introduces the
concept of post translators. Post translators are applied to a job after
Zotero's translators and they can add additional meta data to a job.
This commit implements a post translator for arXiv that reads out the
"Journal reference" field on the web page, if existent. We still enforce
manual croping for arXiv but we use the information from the "Journal
reference" field on arXiv to initialize the crop window with
better default settings.