New database schema
Use a new layout of database tables. The new schema allows for more specific
filters (most importantly: minDate and maxDate). It also allows to set
specific k2pdfopt command line parameters on a per-journal level and it
defines precedence of conflicting filters and makes it possible to
overwrite built-in filters in a well-defined way both by the user
and by updates.

This commit includes
  * Creation of the SQL tables.
  * Port of pdfcrop (including revised UI)
  * Port from inches to mm

The following is separated into the next commits:
  * Port of convertDocument()
  * Port of reportjournalSettings()
  * Implementation of a new tab on the options window
    that allows to edit the journal settings
  * Implementation of privacy settings (neverRoportJournalSettings)