Simplify updateDontprintIconVisibility()
There are three different menu items labeled "Dontprint this page" which
have to be set enabled or disabled depending on the currently displayed
page: In "Tools --> Dontprint" on Linux & Mac, in the minified Windows
menu and in the menu associated to the toolbar button. Rather than trying
to set the disabled attribute of each of these whenever a new page is
loaded, this commit sets the disabled attribute lazily whenever any
of the menus is about to be displayed. This reduces the code that's
executed on each page load and also allows to reuse code.