Settings for transferMethod="directory" on welcome page.
This implements a part of #1. The welcome page is the page that is
presented when the user restarts Firefox for the first time after installing
Dontprint. This commit adds some UI to the welcome page to set  the
transfer method and set the settings if the transfer method is "directory".
We use Mozilla Snyc to synchronize the transfer method and the settings
specific to the transfer method "email" but not the settings specific for the
transfer method "directory" because users may want to use different target
directories, etc., on different machines. Thus, step 3 of the welcome page
will not show up if the transfer method has already been set to "email" due
to sync but it will show up if the transfer method has been set to "directory"
due to sync.