Implement the program logic for transferMethod==="directory"
This implements the actual program logic behind #1. If the user chooses
the transferMethod "directory" in the settings, then Dontprint now first
writes the output of k2pdfopt to some temporary file, which is then copied
to the directory requested by the user. This detour minimizes the duration
for which the destination directory is written to, which may be more
convenient if the destination directory is on a removable file system.

User feedback is not perfect yet. In particular, the job is aborted and
cleaned up if the move operation fails. It would be better to ask the user for
an alternative directory where the file should be copied in case the original
choice didn't work out. Also, the progress page (job queue) needs to be
adapted to this new kind of jobs.