Update progress page for jobs with transferMethod==="directory"
For jobs with transferMethod==="directory", the two tasks "upload" and
"send" in the progress bar on the job queue make no sense. Instead, they
have a "transfer" and a "notify" task. The former means that the file is
being moved to the destination directory (which may be a long-running
operation if the destination directory is on a remote device) while the latter
means that the user-defined post-process command is being executed.

Resolves #1.