Increase minimum required Firefox version to 21.0
According to a user's report, the Dontprint welcome screen does not show up on Firefox 20. The same user reports that Dontprint does work on the latest version of Firefox (version 27) on the same system. Since Firefox 20 is no longer supported anyway, it makes no sense to backport Dontprint to that version.