Fix bug where settings are not saved on Linux+Mac
When the preference window is closed by clicking the 'x' on
the window frame, the "dialogaccept" event does not fire
for whatever reason. Therefore, the currently edited journal
filter is not being saved. This commit fixes this issue by
listening additionally to the "close" handler. Note that we
still need to listen to the "dialogaccept" event, too, as
the "close" event does not seem to fire if the user closes
the preference window by hitting the enter key while the
cursor is in a textbox (at least not on Linux).