Fix bug with editing builtin journal filters.
When the user edits a builtin journal filter, the builtin
filter gets disabled and a new custom filter with the edited
values is added. However, due to a bug in sqlite's lastInsertRowID,
if the edited filter was deselected, then selected again and edited
again, *another* custom filter used to be added. This commit fixes
this issue by replacing the lookup of lastInsertRowID by something
more stable.

Additionally, there was a bug where the id of the first custom filter
(id=0) clashed with the marker that marked a new (not yet saved in
database) filter. We now mark new filters with id===undefined instead
of id=0.