Add "Share on Facebook" button to result page
Since the result page is now loaded from the web, this
commit does not actually contain the result page but it
does contain some changes to the backend-code that were
necessary to make this work. In particular, we now extract
the author names from the zotero item and store them in
the job object.