Add Dontprint button to fullscreen Zotero
Add an overlay to Zotero's tab.xul, which is displayed when
the user clicks the "Tab mode" button in the Zotero pane.
The overlay adds a Dontprint button to the Zotero tab that
is functionally equivalent to the Dontprint button that
already exists on the Zotero pane. On initialization, the
Dontprint button on the Zotero tab is being registered with
the Dontprint browser so that its appearance can be changed
even if the tab is not being xdisplayed in the foreground.

Fixes #4.