Set correct "title" meta-data of ouput PDF
This is important because Kobo e-readers apparently display the
title meta-data instead of the file name. We set the title field
by using already the finally intended file name for the temporarily
created file. It seems like k2pdfopt sets the title meta-data
to the output file name, at least if the input file did not already
hava a title field.

Fixes #15.