Fix severe issue with long file paths on Mac OS X
Fixes a bug introduced in commit 183d3ba. On Mac OS X, k2pdfopt
shortens long file paths for some reason. The effect of this was
that k2pdfopt generated a file with an unexpected filename if
the requested file path was too long. We work around this by
(1) using a shorter filename for the (temporary) output file of
k2pdfopt and (2) we run k2pdfopt from the directory of the
output file so that we can refer to it by relative path and
don't need to include the whole path in the command line
argument. Measure (1) does not affect the final name of the
file after the transfer/email step but it does affect the meta-
data that is included in the PDF document.