Pragmatic fix to issue with pref-window resizing on Mac
Fixes #16.

In fact, this just makes the initial size of the iframe slightly
bigger. An actual dynamic fix seems difficult (as indicated by
reports of frustration with simmilar issues on stackexchange).
It's related with browser.preferences.animateFadeIn being set to
true on mac. This flag seems to generate a lot of problems,
also for other projects.