Exclude some types of webpages to be detected by Dontprint
Dontprint used to display a "dontprint" icon on a lot of pages
that didn't actually correspond to scientific journal articles.
When clicking the dontprint icon, this lead to an error message
scince Dontprint couldn't find any corresponding PDF document.

We now filter out the most common ones of these false positives
by blacklisting some Zotero "itemTypes". The blacklist is rather
short because, when in doubt, we still prefer displaying a
dontprint icon on a page where it isn't useful ("false positives")
rather than displaying no icon on a page where it would be
useful ("false negatives"). Thus, we blacklist only itemTypes
that are both common and clearly not useful.

One effect of this change is that Firefox will no longer show
a dontprint icon on the website www.dontprint.net. This is a
desired effect.