Collect names of non-kindle ereaders
If the user selects "Not a Kindle" on the "device" tab of the preferences
window and checks the box to report the settings back to me, then dontprint
now shows an input box where the user is asked to enter the model name.
This information is transmitted alongside the screen settings as soon as
the user closes the preferences window.

Fixes #7.