Fix conflicts with zotero addon
The Dontprint addon sometimes interfered with the Zotero addon. The
reaseon was that Dontprint tried to get a reference to the Zotero
service while the Dontprint service was registered. This failed if
the Zotero service was registered after Dontprint. In this case, it
also seemed to pollute the Zotero namespace causing Zotero to crash
when trying to display the Zotero pane.

We now call getService() on the Zotero service only from js-code that
is included in xul pages, and therefore executed *after* all services
are registered.

Also fixes #6.