Improve initialization of sqlite database
This fixes a mistake in commit 5a7a2b0a2. Note that, while the
code in commit 5a7a2b0a2 had a bug, it didn't cause any failure
because the inncorrect statement was in a "try" block. In the end,
the code in commit 5a7a2b0a2 still did the correct thing. Thus,
there is no need to publish another bugfix version of Dontprint
just to include this commit.