Remove HTTP requests to zotero server
The Zotero Connector, on which Dontprint for Chrome is based, used
to download its translators from a Zotero-run server *every* Browser
session. This commit removes these requests and packages all translators
as part of the Browser extension bundle. The packaging of all translators
is done automatically by the build script. For further explanations, see
commit message in changes sub-sub-module
/dependencies/zotero-connectors/src/zotero