Update builtin zotero to version 4.0.28.7 and fix tranlsation on FF
On firefox, the new (portable) version of Dontprint was only able to
run Zotero translators using the builtin Zotero. If the Zotero addon
was installed on Firefox, website detection failed. This was because
the builtin Zotero version was outdated and the latest Zotero version
had changed the way translators were associated with tabs. This
commit updates the builtin Zotero to version 4.0.28.7 (with some
Dontprint-related changes) and adapts Dontprint's website detection
and translation to this version of Zotero. Tested on Chrome and on
Firefox with and without Zotero installed.