Blacklist some item types that cannot be handled by Dontprint
This is to port commit 0bf04cf to the portable version of Dontprint.

Also improved the build script for Dontprint on chrome and added a
method "Dontprint.aboutDontprint()" that can be called, e.g., from
the dontprint.net website to detect if Dontprint is already installed
and up to date.