On Chrome, load k2pdfopt NaCl module once after installation/update
When the k2pdfopt NaCl module is loaded for the first time, this takes
a lot of time (presumably because the portable NaCl has to be
translated into architecture-specific NaCl once, which will then be
chached for subsequent calls). For this reason, we now listen to the
chrome.runtime.onInstalled event, which is fired when Dontprint is
installed or updated. When the event is fired, we load k2pdfopt once
and run a dummy command that gets the k2pdfopt version.