Add title and author meta data to optimized PDF also on Firefox
K2pdfopt version 2.33 introduced the "-author" and "-title" command
line arguments with which one can set the author and title meta data
of the ouput PDF. If k2pdfopt version 2.33 or newer is installed,
Dontprint now uses these command line arguments on Firefox.

On Google Chrome, we used to pass the author and title via a special
set of attributes in the NaCl call. This worked because the NaCl port
of k2pdfopt was hacked to accept author and title meta data in a non-
standard way. With this commit, we update the NaCl port of k2pdfopt
to version 2.33a and remove the hacked title/author attributes in
favour of the ones that are now builtin to upstream k2pdfopt.