Bump version number to v1.1.2 (Firefox only)
This version fixes a bug that was only relevant for Dontprint on
Firefox. Therefore, this version will not be published for
Dontprint on Chrome.

Note that there was never a commit for version 1.1.1 since that
version did not contain any changes. It was only necessary to
bump the version number in order to migrate to Mozilla's new
add-on signing policy.