Let the user change the article title from the cropping tab
In commit 50ac068ac, a text box for the output file name was added to
the cropping tab. This commit replaces this text box by a text box for
the article title. This is a better choice since some e-readers don't
display the file name but the article title instead, and the file name
contains the article title anyway.

The cropping tab is only displayed if Dontprint doesn't know already
know how to crop the pages. It will always be displayed if there's no
meta data available for the article (e.g., if the used clicked on the
Dontprint icon while the article was displayed in the browser's
buildint PDF viewer). This coincides with the most common case when
users would probably want to set the article title manually.

Resolves #34.